repo/pyproject.toml
[build-system]
requires = ["setuptools==69.5.1", "wheel==0.43.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bao-overlap"
version = "0.1.0"
description = "Reproducible pipeline for environment-overlap BAO analysis"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "MSOUP Team"}]
dependencies = [
  "numpy==1.26.4",
  "scipy==1.12.0",
  "pandas==2.2.2",
  "pyyaml==6.0.1",
  "astropy==6.0.1",
  "pyarrow==15.0.2",
  "matplotlib==3.8.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
testpaths = ["tests"]
